requirements.txt
discord.py==1.2.5
Flask
jishaku
pyaml
praw==6.4.0
gunicorn==19.7.1
opencv-python
PILLOW
redis